+++ imaprelay/CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG
 
+## v0.1.2 - 2022-01-09
+  * change "via forwarded" to sender to avoid DMARC issues
+  * check configured folders
+
 ## v0.1.1 - 2021-04-18
   * initial release